+++ SwiftUI_Pull_to_Refresh/Views/City/Weather/SummaryView.swift
@@ -11,6 +11,7 @@ import SwiftUI
 struct SummaryView: View {
 	
 	let celsius: String
+	@Environment(\.citiesFrame) var citiesFrame: CGRect
 	
 	var body: some View {
 		HStack {
@@ -35,5 +36,25 @@ struct SummaryView: View {
 		.padding(.vertical, 20)
 		.padding(.horizontal, 28)
 		.featuredBackgroundStyle()
+		.background(
+			GeometryReader { geometry in
+				background(geometry: geometry)
+			}
+		)
+		.cornerRadius(UI.cornerRadius)
+	}
+	
+	private func background(geometry: GeometryProxy) -> some View {
+		UI.Color.background
+			.overlay(
+				UI.Image.background
+					.backgroundStyle()
+					.background(UI.Color.background)
+					.blur(radius: UI.Image.blur)
+					.offset(
+						x: -geometry.frame(in: .global).origin.x + citiesFrame.origin.x + UI.Spacing.screen,
+						y: -geometry.frame(in: .global).origin.y + citiesFrame.origin.y
+					), alignment: .top
+			)
 	}
 }